[#23] Move raw data urls, functions cleanup
